Move bodhi-client in its own directory
