--- CLIENT_MANIFEST.in
@@ -1,1 +1,0 @@
-include COPYING *.txt
