+++ bodhi-client/bodhi-complete.sh
