+++ bodhi-client/setup.cfg
@@ -1,0 +1,36 @@
+[metadata]
+name = bodhi-client
+version = attr: bodhi.client.__version__
+description = Bodhi client
+#long_description = file: README.rst
+maintainer = Fedora Infrastructure Team
+maintainer_email = infrastructure@lists.fedoraproject.org
+platforms = Fedora, GNU/Linux
+url = https://github.com/fedora-infra/bodhi
+keywords = fedora
+license = GPLv2+
+# Possible options are at https://pypi.python.org/pypi?%3Aaction=list_classifiers
+classifiers =
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Developers
+    Intended Audience :: System Administrators
+    License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)
+    Operating System :: POSIX :: Linux
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Topic :: System :: Software Distribution
+
+[options]
+zip_safe = False
+include_package_data = False
+packages = find_namespace:
+namespace_packages = bodhi
+install_requires =
+    click
+    python-fedora >= 0.9.0
+    koji
+
+[options.entry_points]
+console_scripts =
+    bodhi = bodhi.client:cli
