+++ bodhi-client/tests/fixtures.py
